+++ libs/ngx-editorjs/src/lib/ngx-editorjs.module.ts
@@ -4,14 +4,15 @@ import { FormsModule } from '@angular/forms';
 import { MatFormFieldModule, MatInputModule } from '@angular/material';
 import EditorJS from '@editorjs/editorjs';
 import { createModuleConfig } from './config/module-config';
-import { EditorJSBaseComponent } from './containers/base/container.class';
+import { NgxEditorJSBaseComponent } from './containers/base/container.class';
 import { NgxEditorJSMatFieldComponent } from './containers/editorjs-mat-field/editorjs-mat-field.component';
 import { NgxEditorJSComponent } from './containers/editorjs/editorjs.component';
 import { NgxEditorJSDirective } from './directives/ngx-editorjs.directive';
-import { EditorJSInstance, EDITORJS_MODULE_IMPORT, NgxEditorJSInstanceService } from './services/editorjs-injector';
+import { NgxEditorJSInstanceService } from './services/editorjs-injector';
 import { NgxEditorJSService } from './services/editorjs.service';
 import { NgxEditorJSPluginService } from './services/plugins.service';
 import { FOR_ROOT_OPTIONS_TOKEN, NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './types/config';
+import { EDITORJS_MODULE_IMPORT, EditorJSInstance } from './types/injector';
 
 /**
  * Factory function to return the EditorJS base class
@@ -30,8 +31,8 @@ export function createEditorJSInstance(editorjs: any) {
  */
 @NgModule({
   imports: [CommonModule, FormsModule, MatInputModule, MatFormFieldModule],
-  declarations: [NgxEditorJSComponent, NgxEditorJSMatFieldComponent, NgxEditorJSDirective, EditorJSBaseComponent],
-  exports: [NgxEditorJSComponent, NgxEditorJSMatFieldComponent, NgxEditorJSDirective, EditorJSBaseComponent],
+  declarations: [NgxEditorJSComponent, NgxEditorJSMatFieldComponent, NgxEditorJSDirective, NgxEditorJSBaseComponent],
+  exports: [NgxEditorJSComponent, NgxEditorJSMatFieldComponent, NgxEditorJSDirective, NgxEditorJSBaseComponent],
   providers: [
     NgxEditorJSService,
     NgxEditorJSPluginService,